+++ src/options.ts
@@ -1,20 +1,24 @@
-import { DeltaHandlerFunction, IDeltaHandlerConfigMap } from './deltaHandlers/delta'
 import { IllusoryElement } from './IllusoryElement'
-import { FilterFunction, CloneProcessorFunction } from './utils/duplicateNode'
+import { CloneProcessorFunction, FilterFunction } from './utils/duplicateNode'
 
-export interface IOptions {
+export interface IIllusoryElementOptions {
   /**
-   * If `true`, a deep clone is performed
-   * so that the children are included in the animation
+   * By default, illusory creates a deep clone of each element.
+   * When setting this to `false`, illusory will only perform a shallow clone.
+   *
+   * *Example:*
+   *
+   * `includeChildren: true`
+   *
+   * ![](https://github.com/justintaddei/illusory/blob/assets/includeChildren_true.gif?raw=true)
+   *
+   * `includeChildren: false`
+   *
+   * ![](https://github.com/justintaddei/illusory/blob/assets/includeChildren_false.gif?raw=true)
+   *
    * @default true
    */
   includeChildren: boolean
-  /**
-   * If `true`, only compositor-friendly properties will be animated
-   * (`transform` and `opacity`)
-   * @default false
-   */
-  compositeOnly: boolean
   /**
    * If `false` and the element we're transitioning **to** has a transparent background then
    * the element we're transitioning from will fade out.
@@ -24,6 +28,56 @@ export interface IOptions {
    * @default ['img']
    */
   ignoreTransparency: boolean | string[]
+  /**
+   * If `false` all `data-*` attributes are removed.
+   * Can also be a function that returns `true` if the attribute should be preserved.
+   *
+   * All `data-illusory-*` attributes are always preserved.
+   *
+   * @default false
+   */
+  preserveDataAttributes?: boolean | FilterFunction
+  /**
+   * Called while creating the clone, and then for every ChildNode of
+   * the clone (if `includeChildren: true`). This happens before the `beforeAttach` hook is called.
+   *
+   * @example
+   *  new IllusoryElement(el, {
+   *    processClone(node, depth) {
+   *      if (depth > 0) // Make sure this isn't the root node (e.i. the clone of `el`)
+   *        if (node.tagName === 'VIDEO') // Remove any <video> elements from the clone
+   *          return null
+   *    }
+   *  })
+   */
+  processClone?: CloneProcessorFunction
+  /**
+   * If `true`, `this.attach()` is invoked as soon as the instance is created.
+   * @default false
+   */
+  attachImmediately?: boolean
+}
+
+export interface IIllusoryOptions {
+  /**
+   * Options pertaining to `IllusoryElement`s
+   */
+  element: IIllusoryElementOptions
+  /**
+   * By default, illusory will animate `transform`, `opacity`,
+   * and `border-radius` (`background` is not animated.
+   * Rather, the two elements are cross-faded to give the illusion of it being animated).
+   * Setting `compositeOnly` to `true` will disable the `border-radius` animation, leaving only `transform` and `opacity`.
+   *
+   * *Example:*
+   *
+   * `compositeOnly: true`
+   *
+   * ![](https://github.com/justintaddei/illusory/blob/assets/compositeOnly.gif?raw=true)
+   *
+   * @default false
+   */
+  compositeOnly: boolean
   /**
    * A CSS `<time>`. e.g. `2s`, `150ms`, etc.
    * @default "300ms"
@@ -39,25 +93,39 @@ export interface IOptions {
    * @default 1
    */
   zIndex: number
-  deltaHandlers?: IDeltaHandlerConfigMap
-  beforeAttach?: (from: IllusoryElement, to: IllusoryElement) => void | Promise<void>
-  beforeAnimate?: (from: IllusoryElement, to: IllusoryElement) => void | Promise<void>
-  beforeDetach?: (from: IllusoryElement, to: IllusoryElement) => void | Promise<void>
   /**
-   * If `false` all `data-*` attributes are removed.
-   * Can also be a function that returns `true` if the attribute should be preserved.
-   * All `data-illusory-*` attributes are always preserved.
-   * @default false
+   * Called after the clone is appended to the DOM and the natural element has been
+   * hidden, but before the animation begins.
+   *
+   * @remarks If you return a promise, illusory will wait for the promise to resolve.
    */
-  preserveDataAttributes?: boolean | FilterFunction
-  processClone?: CloneProcessorFunction
+  beforeAnimate?: (from: IllusoryElement, to: IllusoryElement) => void | Promise<unknown>
+  /**
+   * Called after the animation is completed, but before the clone is removed from the DOM.
+   *
+   * @remarks If you return a promise, illusory will wait for the promise to resolve.
+   */
+  beforeDetach?: (from: IllusoryElement, to: IllusoryElement, canceled: boolean) => void | Promise<unknown>
+  /**
+   * An array of scrollable elements (including `document`).
+   * Illusory will listen to `scroll` events on these targets and update the position of the
+   * `IllusoryElement`s so that they appear to remain relative to the given container.
+   *
+   * @tip specifying an empty array (`[]`) will cause the `IllusoryElement`s to remain fixed in the viewport.
+   *
+   * @default [document]
+   */
+  relativeTo: (HTMLElement | Document)[]
 }
 
-export const DEFAULT_OPTIONS: IOptions = {
-  includeChildren: true,
-  ignoreTransparency: ['img'],
+export const DEFAULT_OPTIONS: IIllusoryOptions = {
+  element: {
+    includeChildren: true,
+    ignoreTransparency: ['img']
+  },
+  zIndex: 1,
+  compositeOnly: false,
   duration: '300ms',
   easing: 'ease',
-  zIndex: 1,
-  compositeOnly: false
+  relativeTo: [document]
 }