+++ src/polyfill/NodeList.forEach.ts
@@ -1,9 +1,0 @@
-// From https://developer.mozilla.org/en-US/docs/Web/API/NodeList/forEach
-if (window.NodeList && !NodeList.prototype.forEach) {
-  NodeList.prototype.forEach = function(callback, thisArg) {
-    thisArg = thisArg || window
-    for (let i = 0; i < this.length; i++) {
-      callback.call(thisArg, this[i], i, this)
-    }
-  }
-}